afpd: use fixed-width AFP text encoding hints
AFPName carries a four-byte, big-endian text encoding hint. Restore
v_kTextEncoding to uint32_t (which was mistakenly reverted in commit
d6224b7) and use the same fixed-width type in the charset registry,
removing architecture-dependent long storage and narrowing casts.

Expand kTextEncoding_t to cover the complete AFP 3.4 text encoding
identifier list. Document that defining an identifier does not imply
that Netatalk implements a converter for it, and make the composed
UTF-8 constant explicitly uint32_t.

Change set_name()'s final argument to bool because it selects legacy
versus UTF-8 name serialization; it never supplies the encoding value.
Update callers to pass true or false and continue copying the
volume's cached network-order hint for AFPName output.

Convert UTF-8 encoding hints with htonl() in the AFP test packet
helpers so their generated AFPNames are big-endian on little-endian
hosts.

+++ include/atalk/volume.h
@@ -43,8 +43,7 @@ struct vol {
     charset_t       v_maccharset;
     uint16_t        v_mtou_flags;    /*!< flags for convert_charset in mtoupath */
     uint16_t        v_utom_flags;
-    /* FIXME: should be a u_int32_t ? */
-    long            v_kTextEncoding; /*!< mac charset encoding in network order */
+    uint32_t        v_kTextEncoding; /*!< AFP text encoding hint in network order */
     size_t          max_filename;
     char            *v_veto;
     int             v_adouble;    /*!< adouble format: v1, v2 */
